refactor: auth mock API 책임 분리
