+++ src/features/auth/mock/constants.ts
@@ -1,0 +1,18 @@
+export const MOCK_AUTH_DELAY_MS = 350
+export const MOCK_NICKNAME_CHECK_DELAY_MS = 0
+
+export const TAKEN_NICKNAMES_STORAGE_KEY = 'marble-pop-mock-taken-nicknames'
+export const KAKAO_USER_STORAGE_KEY = 'marble-pop-mock-kakao-user'
+
+export const DEFAULT_TAKEN_NICKNAMES = [
+  '마블왕',
+  '주사위마스터',
+  '행운의여신',
+  '부동산왕',
+  'GoormEE',
+]
+
+export const NICKNAME_PATTERN = /^[A-Za-z0-9가-힣]{2,10}$/
+export const DEFAULT_MOCK_KAKAO_PROFILE_IMAGE =
+  'https://picsum.photos/seed/kakao/96/96'
+export const DEFAULT_MOCK_KAKAO_ID = 'mock-kakao-id-001'
